Old Code cleanup and fixing pf1e Issues
