+++ src/ceus_ProviderEngine.js
@@ -1,3 +1,5 @@
+import { ceus_RollProvider_pf1 } from "./ceus_RollProvider_pf1.js";
+import { ceus_RollProvider_sf1e } from "./ceus_RollProvider_sf1e.js";
 /*import { ceus_RollProvider_pf2e } from "./ceus_RollProvider_pf2e.js";
 import { ceus_RollProvider_cof } from "./ceus_RollProvider_cof.js";
 import { ceus_RollProvider_coc } from "./ceus_RollProvider_coc.js";
@@ -8,13 +10,13 @@ import { ceus_RollProvider_dnd5eJP } from "./ceus_RollProvider_dnd5eJP.js";
 import { ceus_RollProvider_dnd35 } from "./ceus_RollProvider_dnd35.js";
 import { ceus_RollProvider_ffd20 } from "./ceus_RollProvider_ffd20.js";
 import { ceus_RollProvider_ose } from "./ceus_RollProvider_ose.js";
-import { ceus_RollProvider_pf1 } from "./ceus_RollProvider_pf1.js";*/
-import { ceus_RollProvider_sf1e } from "./ceus_RollProvider_sf1e.js";
-//import { ceus_RollProvider_sw5e } from "./ceus_RollProvider_sw5e.js";
+import { ceus_RollProvider_sw5e } from "./ceus_RollProvider_sw5e.js";*/
 
 export class ceus_ProviderEngine { 
 	constructor() {
 		this.externalRollProviders = [
+			new ceus_RollProvider_pf1(),
+			new ceus_RollProvider_sf1e()
             /*new ceus_RollProvider_cd(),
             new ceus_RollProvider_coc(),
             new ceus_RollProvider_cof(),
@@ -24,9 +26,7 @@ export class ceus_ProviderEngine {
             new ceus_RollProvider_dnd35(),
             new ceus_RollProvider_ffd20(),
             new ceus_RollProvider_ose(),
-            new ceus_RollProvider_pf1(),
-            new ceus_RollProvider_pf2e(),*/
-            new ceus_RollProvider_sf1e()/*,
+            new ceus_RollProvider_pf2e(),
             new ceus_RollProvider_sw5e()*/
         ];
 	}
