+++ src/old code/ceus_RollProvider.js
