+++ src/old code/ceus_RollProvider_cd.js
