+++ src/old code/ceus_RollProvider_coc.js
