+++ src/old code/ceus_RollProvider_cof.js
