+++ src/old code/ceus_RollProvider_degenesis.js
