+++ src/old code/ceus_RollProvider_demonlord.js
