+++ src/old code/ceus_RollProvider_dnd5e.js
