+++ src/old code/ceus_RollProvider_ffd20.js
