+++ src/old code/ceus_RollProvider_ose.js
