+++ src/old code/ceus_RollProvider_sw5e.js
